Refactor table-of-contents component to use filtered collection for each output type
