+++ packages/11ty/_includes/components/table-of-contents/epub.js
@@ -1,0 +1,14 @@
+/**
+ * Render `table-of-contents` component with data from collections.epub
+ */
+module.exports = function(eleventyConfig) {
+  const tableOfContentsList = eleventyConfig.getFilter('tableOfContentsList')
+  return function(params) {
+    const { collections, currentPageUrl, presentation } = params
+    return tableOfContentsList({ 
+      collection: collections.tableOfContentsEpub,
+      currentPageUrl,
+      presentation
+    })
+  }
+}
